feat: Setup NiA Shared Module for KMP
