--- app/src/main/kotlin/com/google/samples/apps/nowinandroid/MainActivityViewModel.kt
@@ -1,82 +1,0 @@
-/*
- * Copyright 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.samples.apps.nowinandroid
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.google.samples.apps.nowinandroid.MainActivityUiState.Loading
-import com.google.samples.apps.nowinandroid.MainActivityUiState.Success
-import com.google.samples.apps.nowinandroid.core.data.repository.UserDataRepository
-import com.google.samples.apps.nowinandroid.core.model.data.DarkThemeConfig
-import com.google.samples.apps.nowinandroid.core.model.data.ThemeBrand
-import com.google.samples.apps.nowinandroid.core.model.data.UserData
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
-
-class MainActivityViewModel(
-    userDataRepository: UserDataRepository,
-) : ViewModel() {
-    val uiState: StateFlow<MainActivityUiState> = userDataRepository.userData.map {
-        Success(it)
-    }.stateIn(
-        scope = viewModelScope,
-        initialValue = Loading,
-        started = SharingStarted.WhileSubscribed(5_000),
-    )
-}
-
-sealed interface MainActivityUiState {
-    data object Loading : MainActivityUiState
-
-    data class Success(val userData: UserData) : MainActivityUiState {
-        override val shouldDisableDynamicTheming = !userData.useDynamicColor
-
-        override val shouldUseAndroidTheme: Boolean = when (userData.themeBrand) {
-            ThemeBrand.DEFAULT -> false
-            ThemeBrand.ANDROID -> true
-        }
-
-        override fun shouldUseDarkTheme(isSystemDarkTheme: Boolean) =
-            when (userData.darkThemeConfig) {
-                DarkThemeConfig.FOLLOW_SYSTEM -> isSystemDarkTheme
-                DarkThemeConfig.LIGHT -> false
-                DarkThemeConfig.DARK -> true
-            }
-    }
-
-    /**
-     * Returns `true` if the state wasn't loaded yet and it should keep showing the splash screen.
-     */
-    fun shouldKeepSplashScreen() = this is Loading
-
-    /**
-     * Returns `true` if the dynamic color is disabled.
-     */
-    val shouldDisableDynamicTheming: Boolean get() = true
-
-    /**
-     * Returns `true` if the Android theme should be used.
-     */
-    val shouldUseAndroidTheme: Boolean get() = false
-
-    /**
-     * Returns `true` if dark theme should be used.
-     */
-    fun shouldUseDarkTheme(isSystemDarkTheme: Boolean) = isSystemDarkTheme
-}
