--- app/src/main/kotlin/com/google/samples/apps/nowinandroid/ui/interests2pane/InterestsListDetailScreen.kt
@@ -1,247 +1,0 @@
-/*
- * Copyright 2024 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.samples.apps.nowinandroid.ui.interests2pane
-
-import androidx.activity.compose.BackHandler
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
-import androidx.compose.material3.LocalMinimumInteractiveComponentSize
-import androidx.compose.material3.VerticalDragHandle
-import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
-import androidx.compose.material3.adaptive.WindowAdaptiveInfo
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
-import androidx.compose.material3.adaptive.layout.AnimatedPane
-import androidx.compose.material3.adaptive.layout.ListDetailPaneScaffoldRole
-import androidx.compose.material3.adaptive.layout.PaneAdaptedValue
-import androidx.compose.material3.adaptive.layout.PaneExpansionAnchor
-import androidx.compose.material3.adaptive.layout.ThreePaneScaffoldDestinationItem
-import androidx.compose.material3.adaptive.layout.calculatePaneScaffoldDirective
-import androidx.compose.material3.adaptive.layout.defaultDragHandleSemantics
-import androidx.compose.material3.adaptive.layout.rememberPaneExpansionState
-import androidx.compose.material3.adaptive.navigation.BackNavigationBehavior
-import androidx.compose.material3.adaptive.navigation.NavigableListDetailPaneScaffold
-import androidx.compose.material3.adaptive.navigation.ThreePaneScaffoldNavigator
-import androidx.compose.material3.adaptive.navigation.ThreePaneScaffoldPredictiveBackHandler
-import androidx.compose.material3.adaptive.navigation.rememberListDetailPaneScaffoldNavigator
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clipToBounds
-import androidx.compose.ui.layout.layout
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.google.samples.apps.nowinandroid.feature.interests.InterestsRoute
-import com.google.samples.apps.nowinandroid.feature.interests.navigation.InterestsRoute
-import com.google.samples.apps.nowinandroid.feature.topic.TopicDetailPlaceholder
-import com.google.samples.apps.nowinandroid.feature.topic.TopicScreen
-import com.google.samples.apps.nowinandroid.feature.topic.navigation.TopicRoute
-import kotlinx.coroutines.launch
-import kotlinx.serialization.Serializable
-import org.koin.compose.viewmodel.koinViewModel
-import org.koin.core.parameter.parametersOf
-import kotlin.math.max
-
-@Serializable
-internal object TopicPlaceholderRoute
-
-fun NavGraphBuilder.interestsListDetailScreen() {
-    composable<InterestsRoute> {
-        InterestsListDetailScreen()
-    }
-}
-
-@Composable
-internal fun InterestsListDetailScreen(
-    viewModel: Interests2PaneViewModel = koinViewModel(),
-    windowAdaptiveInfo: WindowAdaptiveInfo = currentWindowAdaptiveInfo(),
-) {
-    val selectedTopicId by viewModel.selectedTopicId.collectAsStateWithLifecycle()
-    InterestsListDetailScreen(
-        selectedTopicId = selectedTopicId,
-        onTopicClick = viewModel::onTopicClick,
-        windowAdaptiveInfo = windowAdaptiveInfo,
-    )
-}
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-@Composable
-internal fun InterestsListDetailScreen(
-    selectedTopicId: String?,
-    onTopicClick: (String) -> Unit,
-    windowAdaptiveInfo: WindowAdaptiveInfo,
-) {
-    val listDetailNavigator = rememberListDetailPaneScaffoldNavigator(
-        scaffoldDirective = calculatePaneScaffoldDirective(windowAdaptiveInfo),
-        initialDestinationHistory = listOfNotNull(
-            ThreePaneScaffoldDestinationItem(ListDetailPaneScaffoldRole.List),
-            ThreePaneScaffoldDestinationItem<Nothing>(ListDetailPaneScaffoldRole.Detail).takeIf {
-                selectedTopicId != null
-            },
-        ),
-    )
-    val coroutineScope = rememberCoroutineScope()
-
-    val paneExpansionState = rememberPaneExpansionState(
-        anchors = listOf(
-            PaneExpansionAnchor.Proportion(0f),
-            PaneExpansionAnchor.Proportion(0.5f),
-            PaneExpansionAnchor.Proportion(1f),
-        ),
-    )
-
-    ThreePaneScaffoldPredictiveBackHandler(
-        listDetailNavigator,
-        BackNavigationBehavior.PopUntilScaffoldValueChange,
-    )
-    BackHandler(
-        paneExpansionState.currentAnchor == PaneExpansionAnchor.Proportion(0f) &&
-            listDetailNavigator.isListPaneVisible() &&
-            listDetailNavigator.isDetailPaneVisible(),
-    ) {
-        coroutineScope.launch {
-            paneExpansionState.animateTo(PaneExpansionAnchor.Proportion(1f))
-        }
-    }
-
-    var topicRoute by remember {
-        val route = selectedTopicId?.let { TopicRoute(id = it) } ?: TopicPlaceholderRoute
-        mutableStateOf(route)
-    }
-
-    fun onTopicClickShowDetailPane(topicId: String) {
-        onTopicClick(topicId)
-        topicRoute = TopicRoute(id = topicId)
-        coroutineScope.launch {
-            listDetailNavigator.navigateTo(ListDetailPaneScaffoldRole.Detail)
-        }
-        if (paneExpansionState.currentAnchor == PaneExpansionAnchor.Proportion(1f)) {
-            coroutineScope.launch {
-                paneExpansionState.animateTo(PaneExpansionAnchor.Proportion(0f))
-            }
-        }
-    }
-
-    val mutableInteractionSource = remember { MutableInteractionSource() }
-    val minPaneWidth = 300.dp
-
-    NavigableListDetailPaneScaffold(
-        navigator = listDetailNavigator,
-        listPane = {
-            AnimatedPane {
-                Box(
-                    modifier = Modifier
-                        .clipToBounds()
-                        .layout { measurable, constraints ->
-                            val width = max(minPaneWidth.roundToPx(), constraints.maxWidth)
-                            val placeable = measurable.measure(
-                                constraints.copy(
-                                    minWidth = minPaneWidth.roundToPx(),
-                                    maxWidth = width,
-                                ),
-                            )
-                            layout(constraints.maxWidth, placeable.height) {
-                                placeable.placeRelative(
-                                    x = 0,
-                                    y = 0,
-                                )
-                            }
-                        },
-                ) {
-                    InterestsRoute(
-                        onTopicClick = ::onTopicClickShowDetailPane,
-                        shouldHighlightSelectedTopic = listDetailNavigator.isDetailPaneVisible(),
-                    )
-                }
-            }
-        },
-        detailPane = {
-            AnimatedPane {
-                Box(
-                    modifier = Modifier
-                        .clipToBounds()
-                        .layout { measurable, constraints ->
-                            val width = max(minPaneWidth.roundToPx(), constraints.maxWidth)
-                            val placeable = measurable.measure(
-                                constraints.copy(
-                                    minWidth = minPaneWidth.roundToPx(),
-                                    maxWidth = width,
-                                ),
-                            )
-                            layout(constraints.maxWidth, placeable.height) {
-                                placeable.placeRelative(
-                                    x = constraints.maxWidth -
-                                        max(constraints.maxWidth, placeable.width),
-                                    y = 0,
-                                )
-                            }
-                        },
-                ) {
-                    AnimatedContent(topicRoute) { route ->
-                        when (route) {
-                            is TopicRoute -> {
-                                TopicScreen(
-                                    showBackButton = !listDetailNavigator.isListPaneVisible(),
-                                    onBackClick = {
-                                        coroutineScope.launch {
-                                            listDetailNavigator.navigateBack()
-                                        }
-                                    },
-                                    onTopicClick = ::onTopicClickShowDetailPane,
-                                    viewModel = koinViewModel(
-                                        key = route.id,
-                                        parameters = { parametersOf(route.id) }
-                                    ),
-                                )
-                            }
-
-                            is TopicPlaceholderRoute -> {
-                                TopicDetailPlaceholder()
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        paneExpansionState = paneExpansionState,
-        paneExpansionDragHandle = {
-            VerticalDragHandle(
-                modifier = Modifier.paneExpansionDraggable(
-                    state = paneExpansionState,
-                    minTouchTargetSize = LocalMinimumInteractiveComponentSize.current,
-                    interactionSource = mutableInteractionSource,
-                    semanticsProperties = paneExpansionState.defaultDragHandleSemantics(),
-                ),
-                interactionSource = mutableInteractionSource,
-            )
-        },
-    )
-}
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private fun <T> ThreePaneScaffoldNavigator<T>.isListPaneVisible(): Boolean =
-    scaffoldValue[ListDetailPaneScaffoldRole.List] == PaneAdaptedValue.Expanded
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private fun <T> ThreePaneScaffoldNavigator<T>.isDetailPaneVisible(): Boolean =
-    scaffoldValue[ListDetailPaneScaffoldRole.Detail] == PaneAdaptedValue.Expanded
