+++ nia-shared/src/commonMain/composeResources/values/strings.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright 2024 Mifos Initiative
+
+    This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+    If a copy of the MPL was not distributed with this file,
+    You can obtain one at https://mozilla.org/MPL/2.0/.
+
+    See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
+-->
+<resources>
+    <string name="app_name">Now in Android</string>
+    <string name="not_connected">⚠️ You aren’t connected to the internet</string>
+    <string name="foryou_title">For you</string>
+    <string name="bookmarks_title">Saved</string>
+    <string name="search_interests">Interests</string>
+</resources>
