+++ nia-shared/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/shared/di/JankStatsModule.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.shared.di
+
+import com.google.samples.apps.nowinandroid.shared.NiaSharedAppViewModel
+import com.google.samples.apps.nowinandroid.shared.ui.interests2pane.Interests2PaneViewModel
+import org.koin.core.module.Module
+import org.koin.core.module.dsl.viewModelOf
+import org.koin.dsl.module
+
+internal val sharedModule = module {
+    includes(jankStatsModule)
+
+    viewModelOf(::NiaSharedAppViewModel)
+    viewModelOf(::Interests2PaneViewModel)
+}
+
+internal expect val jankStatsModule: Module
