+++ nia-shared/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/shared/ui/NiaAppState.kt
@@ -1,0 +1,184 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.shared.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.util.trace
+import androidx.navigation.NavDestination
+import androidx.navigation.NavDestination.Companion.hasRoute
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navOptions
+import com.google.samples.apps.nowinandroid.core.data.repository.UserNewsResourceRepository
+import com.google.samples.apps.nowinandroid.core.data.util.NetworkMonitor
+import com.google.samples.apps.nowinandroid.core.data.util.TimeZoneMonitor
+import com.google.samples.apps.nowinandroid.feature.bookmarks.navigation.navigateToBookmarks
+import com.google.samples.apps.nowinandroid.feature.foryou.navigation.navigateToForYou
+import com.google.samples.apps.nowinandroid.feature.interests.navigation.navigateToInterests
+import com.google.samples.apps.nowinandroid.feature.search.navigation.navigateToSearch
+import com.google.samples.apps.nowinandroid.shared.navigation.TopLevelDestination
+import com.google.samples.apps.nowinandroid.shared.navigation.TopLevelDestination.BOOKMARKS
+import com.google.samples.apps.nowinandroid.shared.navigation.TopLevelDestination.FOR_YOU
+import com.google.samples.apps.nowinandroid.shared.navigation.TopLevelDestination.INTERESTS
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
+import kotlinx.datetime.TimeZone
+
+@Composable
+fun rememberNiaAppState(
+    networkMonitor: NetworkMonitor,
+    userNewsResourceRepository: UserNewsResourceRepository,
+    timeZoneMonitor: TimeZoneMonitor,
+    coroutineScope: CoroutineScope = rememberCoroutineScope(),
+    navController: NavHostController = rememberNavController(),
+): NiaAppState {
+    NavigationTrackingSideEffect(navController)
+    return remember(
+        navController,
+        coroutineScope,
+        networkMonitor,
+        userNewsResourceRepository,
+        timeZoneMonitor,
+    ) {
+        NiaAppState(
+            navController = navController,
+            coroutineScope = coroutineScope,
+            networkMonitor = networkMonitor,
+            userNewsResourceRepository = userNewsResourceRepository,
+            timeZoneMonitor = timeZoneMonitor,
+        )
+    }
+}
+
+@Stable
+class NiaAppState(
+    val navController: NavHostController,
+    coroutineScope: CoroutineScope,
+    networkMonitor: NetworkMonitor,
+    userNewsResourceRepository: UserNewsResourceRepository,
+    timeZoneMonitor: TimeZoneMonitor,
+) {
+    private val previousDestination = mutableStateOf<NavDestination?>(null)
+
+    val currentDestination: NavDestination?
+        @Composable get() {
+            // Collect the currentBackStackEntryFlow as a state
+            val currentEntry = navController.currentBackStackEntryFlow
+                .collectAsState(initial = null)
+
+            // Fallback to previousDestination if currentEntry is null
+            return currentEntry.value?.destination.also { destination ->
+                if (destination != null) {
+                    previousDestination.value = destination
+                }
+            } ?: previousDestination.value
+        }
+
+    val currentTopLevelDestination: TopLevelDestination?
+        @Composable get() {
+            return TopLevelDestination.entries.firstOrNull { topLevelDestination ->
+                currentDestination?.hasRoute(route = topLevelDestination.route) == true
+            }
+        }
+
+    val isOffline = networkMonitor.isOnline
+        .map(Boolean::not)
+        .stateIn(
+            scope = coroutineScope,
+            started = SharingStarted.WhileSubscribed(5_000),
+            initialValue = false,
+        )
+
+    /**
+     * Map of top level destinations to be used in the TopBar, BottomBar and NavRail. The key is the
+     * route.
+     */
+    val topLevelDestinations: List<TopLevelDestination> = TopLevelDestination.entries
+
+    /**
+     * The top level destinations that have unread news resources.
+     */
+    val topLevelDestinationsWithUnreadResources: StateFlow<Set<TopLevelDestination>> =
+        userNewsResourceRepository.observeAllForFollowedTopics()
+            .combine(userNewsResourceRepository.observeAllBookmarked()) { forYouNewsResources, bookmarkedNewsResources ->
+                setOfNotNull(
+                    FOR_YOU.takeIf { forYouNewsResources.any { !it.hasBeenViewed } },
+                    BOOKMARKS.takeIf { bookmarkedNewsResources.any { !it.hasBeenViewed } },
+                )
+            }
+            .stateIn(
+                coroutineScope,
+                SharingStarted.WhileSubscribed(5_000),
+                initialValue = emptySet(),
+            )
+
+    val currentTimeZone = timeZoneMonitor.currentTimeZone
+        .stateIn(
+            coroutineScope,
+            SharingStarted.WhileSubscribed(5_000),
+            TimeZone.currentSystemDefault(),
+        )
+
+    /**
+     * UI logic for navigating to a top level destination in the app. Top level destinations have
+     * only one copy of the destination of the back stack, and save and restore state whenever you
+     * navigate to and from it.
+     *
+     * @param topLevelDestination: The destination the app needs to navigate to.
+     */
+    fun navigateToTopLevelDestination(topLevelDestination: TopLevelDestination) {
+        trace("Navigation: ${topLevelDestination.name}") {
+            val topLevelNavOptions = navOptions {
+                // Pop up to the start destination of the graph to
+                // avoid building up a large stack of destinations
+                // on the back stack as users select items
+                popUpTo(navController.graph.findStartDestination().id) {
+                    saveState = true
+                }
+                // Avoid multiple copies of the same destination when
+                // reselecting the same item
+                launchSingleTop = true
+                // Restore state when reselecting a previously selected item
+                restoreState = true
+            }
+
+            when (topLevelDestination) {
+                FOR_YOU -> navController.navigateToForYou(topLevelNavOptions)
+                BOOKMARKS -> navController.navigateToBookmarks(topLevelNavOptions)
+                INTERESTS -> navController.navigateToInterests(null, topLevelNavOptions)
+            }
+        }
+    }
+
+    fun navigateToSearch() = navController.navigateToSearch()
+}
+
+/**
+ * Stores information about navigation events to be used with JankStats
+ */
+@Composable
+internal expect fun NavigationTrackingSideEffect(navController: NavHostController)
