+++ nia-shared/src/desktopMain/kotlin/com/google/samples/apps/nowinandroid/shared/ui/NiaAppState.desktop.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.shared.ui
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+/**
+ * Stores information about navigation events to be used with JankStats
+ */
+@Composable
+internal actual fun NavigationTrackingSideEffect(navController: NavHostController) {
+    // NO-OP
+}
